backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.46
PyMySQL==1.1.0
cryptography==43.0.0
python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
alembic==1.13.1
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.0
python-dotenv==1.0.0
email-validator>=2.1.0

